requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
google-generativeai==0.3.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.26.4  # Or whatever version pip installs
scikit-learn==1.4.2 # The new version
scipy==1.13.0 # Or whatever version pip installs for scipy
mcp==1.0.0
websockets==12.0